Integer overflow in IndexedDB string key length bypasses bounds check https://bugs.webkit.org/show_bug.cgi?id=308713 rdar://171245941
Reviewed by Sihui Liu.

Make it impossible for the bounds check to overflow.

Tests: Tools/TestWebKitAPI/Tests/WebCore/IDBSerializationTest.cpp
Canonical link: https://commits.webkit.org/308649@main

+++ Source/WebCore/Modules/indexeddb/server/IDBSerialization.cpp
@@ -323,7 +323,7 @@ RefPtr<SharedBuffer> serializeIDBKeyData(const IDBKeyData& key)
         if (!readLittleEndian(data, length))
             return false;
 
-        if (data.size() < length * 2)
+        if (data.size() < static_cast<size_t>(length) * 2)
             return false;
 
         Vector<char16_t> buffer;
